backend/pyproject.toml
[project]
name = "flask-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "pyjwt>=2.10.1",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
